+++ Materials/UnLitTextured/UnLitDescriptorSets.h
@@ -59,7 +59,7 @@ namespace dmbrn
 
 			descriptor_sets_ = device->allocateDescriptorSets(allocInfo);
 
-			for (size_t i = 0; i < device.MAX_FRAMES_IN_FLIGHT; i++)
+			for (uint32_t i = 0; i < device.MAX_FRAMES_IN_FLIGHT; i++)
 			{
 				vk::DescriptorBufferInfo bufferInfo
 				{
